Refactor db initialization
